Refactor base module into pipeline and io
