+++ WhiskiWrap/base.py
@@ -1126,1536 +1126,5 @@ def read_whisker_data(filename, output_format='dict'):
     elif output_format == 'df':
         return meas_table
 
-def pipeline_trace(input_vfile, h5_filename,
-    epoch_sz_frames=3200, chunk_sz_frames=200,
-    frame_start=0, frame_stop=None,
-    n_trace_processes=4, expectedrows=1000000, flush_interval=100000,
-    measure=False,face='right'):
-    """Trace a video file using a chunked strategy.
-
-    This is now deprecated in favor of interleaved_reading_and_tracing.
-    The issue with this function is that it has to write out all of the tiffs
-    first, before tracing, which is a wasteful use of disk space.
-
-    input_vfile : input video filename
-    h5_filename : output HDF5 file
-    epoch_sz_frames : Video is first broken into epochs of this length
-    chunk_sz_frames : Each epoch is broken into chunks of this length
-    frame_start, frame_stop : where to start and stop processing
-    n_trace_processes : how many simultaneous processes to use for tracing
-    expectedrows, flush_interval : used to set up hdf5 file
-
-    TODO: combine the reading and writing stages using frame_func so that
-    we don't have to load the whole epoch in at once. In fact then we don't
-    even need epochs at all.
-    """
-    wwutils.utils.probe_needed_commands()
-
-    # Figure out where to store temporary data
-    input_vfile = os.path.abspath(input_vfile)
-    input_dir = os.path.split(input_vfile)[0]
-
-    # Setup the result file
-    setup_hdf5(h5_filename, expectedrows, measure=measure)
-
-    # Figure out how many frames and epochs
-    duration = video.get_video_duration(input_vfile)
-    frame_rate = video.get_video_params(input_vfile)[2]
-    total_frames = int(np.rint(duration * frame_rate))
-    if frame_stop is None:
-        frame_stop = total_frames
-    if frame_stop > total_frames:
-        print("too many frames requested, truncating")
-        frame_stop = total_frames
-
-    # Iterate over epochs
-    for start_epoch in range(frame_start, frame_stop, epoch_sz_frames):
-        # How many frames in this epoch
-        stop_epoch = np.min([frame_stop, start_epoch + epoch_sz_frames])
-        print(("Epoch %d - %d" % (start_epoch, stop_epoch)))
-
-        # Chunks
-        chunk_starts = np.arange(start_epoch, stop_epoch, chunk_sz_frames)
-        chunk_names = ['chunk%08d.tif' % nframe for nframe in chunk_starts]
-        whisk_names = ['chunk%08d.whiskers' % nframe for nframe in chunk_starts]
-
-        # read everything
-        # need to be able to crop here
-        print("Reading")
-        frames = video.process_chunks_of_video(input_vfile,
-            frame_start=start_epoch, frame_stop=stop_epoch,
-            frames_per_chunk=chunk_sz_frames, # only necessary for chunk_func
-            frame_func=None, chunk_func=None,
-            verbose=False, finalize='listcomp')
-
-        # Dump frames into tiffs or lossless
-        print("Writing")
-        for n_whiski_chunk, chunk_name in enumerate(chunk_names):
-            print(n_whiski_chunk)
-            chunkstart = n_whiski_chunk * chunk_sz_frames
-            chunkstop = (n_whiski_chunk + 1) * chunk_sz_frames
-            chunk = frames[chunkstart:chunkstop]
-            if len(chunk) in [3, 4]:
-                print("WARNING: trace will fail on tiff stacks of length 3 or 4")
-            write_chunk(chunk, chunk_name, input_dir)
-
-        # Also write lossless and/or lossy monitor video here?
-        # would really only be useful if cropping applied
-
-        # trace each
-        print("Tracing")
-        pool = mp.Pool(n_trace_processes)
-        trace_res = pool.map(trace_chunk,
-            [os.path.join(input_dir, chunk_name)
-                for chunk_name in chunk_names])
-        pool.close()
-
-        # take measurements:
-        if measure:
-            print("Measuring")
-            pool = mp.Pool(n_trace_processes)
-            meas_res = pool.map(measure_chunk_star,
-                list(zip([os.path.join(input_dir, whisk_name)
-                    for whisk_name in whisk_names],itertools.repeat(face))))
-            pool.close()
-
-        # stitch
-        print("Stitching")
-        for chunk_start, chunk_name in zip(chunk_starts, chunk_names):
-            # Append each chunk to the hdf5 file
-            fn = wwutils.utils.FileNamer.from_tiff_stack(
-                os.path.join(input_dir, chunk_name))
-
-            if not measure:
-                append_whiskers_to_hdf5(
-                    whisk_filename=fn.whiskers,
-                    h5_filename=h5_filename,
-                    chunk_start=chunk_start)
-            elif measure:
-                append_whiskers_to_hdf5(
-                    whisk_filename=fn.whiskers,
-                    measurements_filename=fn.measurements,
-                    h5_filename=h5_filename,
-                    chunk_start=chunk_start)
-
-def write_video_as_chunked_tiffs(input_reader, tiffs_to_trace_directory,
-    chunk_size=200, chunk_name_pattern='chunk%08d.tif',
-    stop_after_frame=None, monitor_video=None, timestamps_filename=None,
-    monitor_video_kwargs=None):
-    """Write frames to disk as tiff stacks
-
-    input_reader : object providing .iter_frames() method and perhaps
-        also a .timestamps attribute. For instance, PFReader, or some
-        FFmpegReader object.
-    tiffs_to_trace_directory : where to store the chunked tiffs
-    stop_after_frame : to stop early
-    monitor_video : if not None, should be a filename to write a movie to
-    timestamps_filename : if not None, should be the name to write timestamps
-    monitor_video_kwargs : ffmpeg params
-
-    Returns: ChunkedTiffWriter object
-    """
-    # Tiff writer
-    ctw = WhiskiWrap.ChunkedTiffWriter(tiffs_to_trace_directory,
-        chunk_size=chunk_size, chunk_name_pattern=chunk_name_pattern)
-
-    # FFmpeg writer is initalized after first frame
-    ffw = None
-
-    # Iterate over frames
-    for nframe, frame in enumerate(input_reader.iter_frames()):
-        # Stop early?
-        if stop_after_frame is not None and nframe >= stop_after_frame:
-            break
-
-        # Write to chunked tiff
-        ctw.write(frame)
-
-        # Optionally write to monitor video
-        if monitor_video is not None:
-            # Initialize ffw after first frame so we know the size
-            if ffw is None:
-                ffw = WhiskiWrap.FFmpegWriter(monitor_video,
-                    frame_width=frame.shape[1], frame_height=frame.shape[0],
-                    **monitor_video_kwargs)
-            ffw.write(frame)
-
-    # Finalize writers
-    ctw.close()
-    if ffw is not None:
-        ff_stdout, ff_stderr = ffw.close()
-
-    # Also write timestamps as numpy file
-    if hasattr(input_reader, 'timestamps') and timestamps_filename is not None:
-        timestamps = np.concatenate(input_reader.timestamps)
-        assert len(timestamps) >= ctw.frames_written
-        np.save(timestamps_filename, timestamps[:ctw.frames_written])
-
-    return ctw
-
-def trace_chunked_tiffs(input_tiff_directory, h5_filename,
-    n_trace_processes=4, expectedrows=1000000,
-    ):
-    """Trace tiffs that have been written to disk in parallel and stitch.
-
-    input_tiff_directory : directory containing tiffs
-    h5_filename : output HDF5 file
-    n_trace_processes : how many simultaneous processes to use for tracing
-    expectedrows : used to set up hdf5 file
-    """
-    wwutils.utils.probe_needed_commands()
-
-    # Setup the result file
-    setup_hdf5(h5_filename, expectedrows)
-
-    # The tiffs have been written, figure out which they are
-    tif_file_number_strings = wwutils.misc.apply_and_filter_by_regex(
-        '^chunk(\d+).tif$', os.listdir(input_tiff_directory), sort=False)
-    tif_full_filenames = [
-        os.path.join(input_tiff_directory, 'chunk%s.tif' % fns)
-        for fns in tif_file_number_strings]
-    tif_file_numbers = list(map(int, tif_file_number_strings))
-    tif_ordering = np.argsort(tif_file_numbers)
-    tif_sorted_filenames = np.array(tif_full_filenames)[
-        tif_ordering]
-    tif_sorted_file_numbers = np.array(tif_file_numbers)[
-        tif_ordering]
-
-    # trace each
-    print("Tracing")
-    pool = mp.Pool(n_trace_processes)
-    trace_res = pool.map(trace_chunk, tif_sorted_filenames)
-    pool.close()
-
-    # stitch
-    print("Stitching")
-    for chunk_start, chunk_name in zip(tif_sorted_file_numbers, tif_sorted_filenames):
-        # Append each chunk to the hdf5 file
-        fn = wwutils.utils.FileNamer.from_tiff_stack(chunk_name)
-        append_whiskers_to_hdf5(
-            whisk_filename=fn.whiskers,
-            h5_filename=h5_filename,
-            chunk_start=chunk_start)
-
-def interleaved_read_trace_and_measure(input_reader, tiffs_to_trace_directory,
-    sensitive=False,
-    chunk_size=200, chunk_name_pattern='chunk%08d.tif',
-    stop_after_frame=None, delete_tiffs=True,
-    timestamps_filename=None, monitor_video=None,
-    monitor_video_kwargs=None, write_monitor_ffmpeg_stderr_to_screen=False,
-    h5_filename=None, frame_func=None,
-    n_trace_processes=4, expectedrows=1000000,
-    verbose=True, skip_stitch=False, face='right'
-    ):
-    """Read, write, and trace each chunk, one at a time.
-
-    This is an alternative to first calling:
-        write_video_as_chunked_tiffs
-    And then calling
-        trace_chunked_tiffs
-
-    input_reader : Typically a PFReader or FFmpegReader
-    tiffs_to_trace_directory : Location to write the tiffs
-    sensitive: if False, use default. If True, lower MIN_SIGNAL
-    chunk_size : frames per chunk
-    chunk_name_pattern : how to name them
-    stop_after_frame : break early, for debugging
-    delete_tiffs : whether to delete tiffs after done tracing
-    timestamps_filename : Where to store the timestamps
-        Only vallid for PFReader input_reader
-    monitor_video : filename for a monitor video
-        If None, no monitor video will be written
-    monitor_video_kwargs : kwargs to pass to FFmpegWriter for monitor
-    write_monitor_ffmpeg_stderr_to_screen : whether to display
-        output from ffmpeg writing instance
-    h5_filename : hdf5 file to stitch whiskers information into
-    frame_func : function to apply to each frame
-        If 'invert', will apply 255 - frame
-    n_trace_processes : number of simultaneous trace processes
-    expectedrows : how to set up hdf5 file
-    verbose : verbose
-    skip_stitch : skip the stitching phase
-
-    Returns: dict
-        trace_pool_results : result of each call to trace
-        monitor_ff_stderr, monitor_ff_stdout : results from monitor
-            video ffmpeg instance
-    """
-    ## Set up kwargs
-    if monitor_video_kwargs is None:
-        monitor_video_kwargs = {}
-
-    if frame_func == 'invert':
-        frame_func = lambda frame: 255 - frame
-
-    # Check commands
-    wwutils.utils.probe_needed_commands()
-
-    ## Initialize readers and writers
-    if verbose:
-        print("initalizing readers and writers")
-    # Tiff writer
-    ctw = WhiskiWrap.ChunkedTiffWriter(tiffs_to_trace_directory,
-        chunk_size=chunk_size, chunk_name_pattern=chunk_name_pattern)
-
-    # FFmpeg writer is initalized after first frame
-    ffw = None
-
-    # Setup the result file
-    setup_hdf5(h5_filename, expectedrows, measure=True)
-
-    # Copy the parameters files
-    copy_parameters_files(tiffs_to_trace_directory, sensitive=sensitive)
-
-    ## Set up the worker pool
-    # Pool of trace workers
-    trace_pool = mp.Pool(n_trace_processes)
-
-    # Keep track of results
-    trace_pool_results = []
-    # deleted_tiffs = []
-    def log_result(result):
-        print("Result logged:", result)  # Verify the callback
-        trace_pool_results.append(result)
-
-    ## Iterate over chunks
-    out_of_frames = False
-    nframe = 0
-
-    # Init the iterator outside of the loop so that it persists
-    iter_obj = input_reader.iter_frames()
-
-    while not out_of_frames:
-        # Get a chunk of frames
-        if verbose:
-            print("loading chunk of frames starting with", nframe)
-        chunk_of_frames = []
-        for frame in iter_obj:
-            if frame_func is not None:
-                frame = frame_func(frame)
-            chunk_of_frames.append(frame)
-            nframe = nframe + 1
-            if stop_after_frame is not None and nframe >= stop_after_frame:
-                break
-            if len(chunk_of_frames) == chunk_size:
-                break
-
-        # Check if we ran out
-        if len(chunk_of_frames) != chunk_size:
-            out_of_frames = True
-
-        ## Write tiffs
-        # We do this synchronously to ensure that it happens before
-        # the trace starts
-        for frame in chunk_of_frames:
-            ctw.write(frame)
-
-        # Make sure the chunk was written, in case this is the last one
-        # and we didn't reach chunk_size yet
-        if len(chunk_of_frames) != chunk_size:
-            ctw._write_chunk()
-        assert ctw.count_unwritten_frames() == 0
-
-        # Figure out which tiff file was just generated
-        tif_filename = ctw.chunknames_written[-1]
-
-        ## Start trace
-        trace_pool.apply_async(trace_chunk, args=(tif_filename, delete_tiffs),
-            callback=log_result)
-
-        ## Determine whether we can delete any tiffs
-        #~ if delete_tiffs:
-            #~ tiffs_to_delete = [
-                #~ tpres['video_filename'] for tpres in trace_pool_results
-                #~ if tpres['video_filename'] not in deleted_tiffs]
-            #~ for filename in tiffs_to_delete:
-                #~ if verbose:
-                    #~ print "deleting", filename
-                #~ os.remove(filename)
-
-        ## Start monitor encode
-        # This is also synchronous, otherwise the input buffer might fill up
-        if monitor_video is not None:
-            if ffw is None:
-                ffw = WhiskiWrap.FFmpegWriter(monitor_video,
-                    frame_width=frame.shape[1], frame_height=frame.shape[0],
-                    write_stderr_to_screen=write_monitor_ffmpeg_stderr_to_screen,
-                    **monitor_video_kwargs)
-            for frame in chunk_of_frames:
-                ffw.write(frame)
-
-        ## Determine if we should pause
-        while len(ctw.chunknames_written) > len(trace_pool_results) + 2 * n_trace_processes:
-            print("waiting for tracing to catch up")
-            time.sleep(10)
-
-    ## Wait for trace to complete
-    if verbose:
-        print("done with reading and writing, just waiting for tracing")
-    # Tell it no more jobs, so close when done
-    trace_pool.close()
-
-    # Wait for everything to finish
-    trace_pool.join()
-
-    ## Error check the tifs that were processed
-    # Get the tifs we wrote, and the tifs we trace
-    written_chunks = sorted(ctw.chunknames_written)
-    traced_filenames = sorted([
-        res['video_filename'] for res in trace_pool_results])
-
-    # Check that they are the same
-    if not np.all(np.array(written_chunks) == np.array(traced_filenames)):
-        raise ValueError("not all chunks were traced")
-
-    ## Extract the chunk numbers from the filenames
-    # The tiffs have been written, figure out which they are
-    split_traced_filenames = [os.path.split(fn)[1] for fn in traced_filenames]
-    tif_file_number_strings = wwutils.misc.apply_and_filter_by_regex(
-        '^chunk(\d+).tif$', split_traced_filenames, sort=False)
-    tif_full_filenames = [
-        os.path.join(tiffs_to_trace_directory, 'chunk%s.tif' % fns)
-        for fns in tif_file_number_strings]
-    tif_file_numbers = list(map(int, tif_file_number_strings))
-    tif_ordering = np.argsort(tif_file_numbers)
-    tif_sorted_filenames = np.array(tif_full_filenames)[
-        tif_ordering]
-    tif_sorted_file_numbers = np.array(tif_file_numbers)[
-        tif_ordering]
-
-    # stitch
-    if not skip_stitch:
-        print("Stitching")
-        zobj = list(zip(tif_sorted_file_numbers, tif_sorted_filenames))
-        for chunk_start, chunk_name in zobj:
-            # Append each chunk to the hdf5 file
-            fn = wwutils.utils.FileNamer.from_tiff_stack(chunk_name)
-            append_whiskers_to_hdf5(
-                whisk_filename=fn.whiskers,
-                h5_filename=h5_filename,
-                chunk_start=chunk_start)
-
-    # Finalize writers
-    ctw.close()
-    if ffw is not None:
-        ff_stdout, ff_stderr = ffw.close()
-    else:
-        ff_stdout, ff_stderr = None, None
-
-    # Also write timestamps as numpy file
-    if hasattr(input_reader, 'timestamps') and timestamps_filename is not None:
-        timestamps = np.concatenate(input_reader.timestamps)
-        assert len(timestamps) >= ctw.frames_written
-        np.save(timestamps_filename, timestamps[:ctw.frames_written])
-
-    return {'trace_pool_results': trace_pool_results,
-        'monitor_ff_stdout': ff_stdout,
-        'monitor_ff_stderr': ff_stderr,
-        'tif_sorted_file_numbers': tif_sorted_file_numbers,
-        'tif_sorted_filenames': tif_sorted_filenames,
-        }
-
-def interleaved_split_trace_and_measure(input_reader, tiffs_to_trace_directory,
-    sensitive=False,
-    chunk_size=200, chunk_name_pattern='chunk%08d.tif',
-    stop_after_frame=None, delete_tiffs=True,
-    timestamps_filename=None, monitor_video=None,
-    monitor_video_kwargs=None, write_monitor_ffmpeg_stderr_to_screen=False,
-    output_filename=None, frame_func=None,
-    n_trace_processes=4, expected_rows=1000000,
-    verbose=True, skip_stitch=False, face='NA', classify=None, 
-    summary_only=False, skip_existing=False, convert_chunks=False
-    ):
-    """Read, write, and trace each chunk, one at a time.
-
-    This is an extension of interleaved_read_trace_and_measure for bilateral whisker tracking.
-    The function needs to be called twice, once for each side of the face. 
-
-    input_reader : Typically a PFReader or FFmpegReader
-    tiffs_to_trace_directory : Location to write the tiffs
-    sensitive: if False, use default. If True, lower MIN_SIGNAL
-    chunk_size : frames per chunk
-    chunk_name_pattern : how to name them
-    stop_after_frame : break early, for debugging
-    delete_tiffs : whether to delete tiffs after done tracing
-    timestamps_filename : Where to store the timestamps
-        Only valid for PFReader input_reader
-    monitor_video : filename for a monitor video
-        If None, no monitor video will be written
-    monitor_video_kwargs : kwargs to pass to FFmpegWriter for monitor
-    write_monitor_ffmpeg_stderr_to_screen : whether to display
-        output from ffmpeg writing instance
-    output_filename : hdf5, zarr or parquet file to stitch whiskers information into
-    frame_func : function to apply to each frame
-        If 'invert', will apply 255 - frame
-    n_trace_processes : number of simultaneous trace processes
-    expected_rows : how to set up hdf5 file
-    verbose : verbose
-    skip_stitch : skip the stitching phase
-    face : 'left','right','top','bottom'
-    classify : if not None, classify whiskers using passed arguments
-
-    Returns: dict
-        trace_pool_results : result of each call to trace
-        monitor_ff_stderr, monitor_ff_stdout : results from monitor
-            video ffmpeg instance
-    """
-    ## Set up kwargs
-    if monitor_video_kwargs is None:
-        monitor_video_kwargs = {}
-
-    if frame_func == 'invert':
-        frame_func = lambda frame: 255 - frame
-
-    if frame_func == 'fliph':
-        frame_func = lambda frame: np.fliplr(frame)
-    elif frame_func == 'flipv':
-        frame_func = lambda frame: np.flipud(frame)
-
-    if frame_func == 'crop':
-        crop_coord = input_reader.crop
-        # crop_coord format is width:height:x:y
-        frame_func = lambda frame: frame[crop_coord[3]:crop_coord[3] + crop_coord[1], crop_coord[2]:crop_coord[2] + crop_coord[0]]
-        # reset crop field to None
-        input_reader.crop = None
-
-    # Check commands
-    wwutils.utils.probe_needed_commands(paths=[whisk_path])
-
-    ## Initialize readers and writers
-    if verbose:
-        print("Initializing readers and writers")
-
-    # Tiff writer
-    ctw = WhiskiWrap.ChunkedTiffWriter(tiffs_to_trace_directory,
-        chunk_size=chunk_size, chunk_name_pattern=chunk_name_pattern)
-
-    # FFmpeg writer is initialized after first frame
-    ffw = None
-
-    # Setup the result file for some formats
-    if not skip_stitch:
-        if output_filename.endswith('.hdf5'):
-            setup_hdf5(output_filename, expected_rows, measure=True)
-        elif output_filename.endswith('.zarr'):
-            initialize_zarr(output_filename, (chunk_size,))
-    
-    # Copy the parameters files
-    copy_parameters_files(tiffs_to_trace_directory, sensitive=sensitive)
-
-    trace_pool_results = []
-
-    if skip_existing:
-    # Check if .whiskers or .measurements files already exist
-        # Remove the extension from the chunk_name_pattern and replace %08d with a regular expression
-        pattern = chunk_name_pattern.replace('.tif', r'(.whiskers|.measurements)').replace('%08d', r'\d{8}')
-        existing_files = [f for f in os.listdir(tiffs_to_trace_directory) if re.match(pattern, f)]
-        # Any whiskers or measurements files (less precise than the pattern):
-        # existing_files = [f for f in os.listdir(tiffs_to_trace_directory) if f.endswith('.whiskers') or f.endsWith('.measurements')]
-
-        if len(existing_files) > 0:    
-            print("Existing files found, skipping to stitching")
-            # create sorted_file_numbers, sorted_filenames
-            # Keep only the .whiskers files in the list
-            split_traced_filenames = [os.path.split(fn)[1] for fn in existing_files]
-            split_traced_filenames = [fn for fn in split_traced_filenames if fn.endswith('.whiskers')]
-            full_filenames = [os.path.join(tiffs_to_trace_directory, fn) for fn in split_traced_filenames]
-            # Get the file numbers
-            chunk_name_pattern = r'_(\d+)\.(whiskers)$'
-            file_number_strings = [re.search(chunk_name_pattern, fn).group(1) 
-                                   for fn in split_traced_filenames if re.search(chunk_name_pattern, fn)]
-            # Sort the filenames by the file numbers
-            file_numbers = list(map(int, file_number_strings))
-            ordering = np.argsort(file_numbers)
-            sorted_filenames = np.array(full_filenames)[ordering]
-            sorted_file_numbers = np.array(file_numbers)[ordering]
-            
-        else: 
-            sorted_filenames = []
-            sorted_file_numbers = []
-            skip_existing = False
-
-    if not skip_existing:
-        ## Set up the worker pool
-        # Pool of trace workers
-        trace_pool = mp.Pool(n_trace_processes)
-
-        # Keep track of results
-        deleted_tiffs = []
-        def log_result(result):
-            print("Result logged:", result)
-            trace_pool_results.append(result)
-
-        ## Iterate over chunks
-        out_of_frames = False
-        nframe = 0
-
-        # Init the iterator outside of the loop so that it persists
-        iter_obj = input_reader.iter_frames()
-
-        # Create temporary directory if output_filename ends with .zarr or. parquet
-        if output_filename.endswith('.zarr') or output_filename.endswith('.parquet'):
-            temp_dir = tempfile.mkdtemp()
-        else:
-            temp_dir = None
-
-        if convert_chunks:
-            #  Check what format to the chunks to based on the output file extension
-            convert_chunks_to = output_filename.split('.')[-1]
-                    
-        while not out_of_frames:
-            # Get a chunk of frames
-            if verbose:
-                print("Loading chunk of frames starting with", nframe)
-            chunk_of_frames = []
-            for frame in iter_obj:
-                if frame_func is not None:
-                    frame = frame_func(frame)
-                chunk_of_frames.append(frame)
-                nframe = nframe + 1
-                if stop_after_frame is not None and nframe >= stop_after_frame:
-                    break
-                if len(chunk_of_frames) == chunk_size:
-                    break
-
-            # Check if we ran out
-            if len(chunk_of_frames) != chunk_size:
-                out_of_frames = True
-
-            ## Write tiffs
-            # We do this synchronously to ensure that it happens before
-            # the trace starts
-            for frame in chunk_of_frames:
-                ctw.write(frame)
-
-            # Make sure the chunk was written, in case this is the last one
-            # and we didn't reach chunk_size yet
-            if len(chunk_of_frames) != chunk_size:
-                ctw._write_chunk()
-            assert ctw.count_unwritten_frames() == 0
-
-            # Figure out which tiff file was just generated
-            tif_filename = ctw.chunknames_written[-1]
-
-            # Start trace
-            try:
-                trace_pool.apply_async(trace_and_measure_chunk, args=(tif_filename, delete_tiffs, face, classify, temp_dir, convert_chunks_to), callback=log_result)
-            except Exception as e:
-                print("Error in apply_async:", e)
-
-            ## Determine whether we can delete any tiffs
-            #~ if delete_tiffs:
-                #~ tiffs_to_delete = [
-                    #~ tpres['video_filename'] for tpres in trace_pool_results
-                    #~ if tpres['video_filename'] not in deleted_tiffs]
-                #~ for filename in tiffs_to_delete:
-                    #~ if verbose:
-                        #~ print "deleting", filename
-                    #~ os.remove(filename)
-
-            ## Start monitor encode
-            # This is also synchronous, otherwise the input buffer might fill up
-            if monitor_video is not None:
-                if ffw is None:
-                    ffw = WhiskiWrap.FFmpegWriter(monitor_video,
-                        frame_width=frame.shape[1], frame_height=frame.shape[0],
-                        write_stderr_to_screen=write_monitor_ffmpeg_stderr_to_screen,
-                        **monitor_video_kwargs)
-                for frame in chunk_of_frames:
-                    ffw.write(frame)
-
-            ## Determine if we should pause
-            # if len(ctw.chunknames_written) > len(trace_pool_results) + 2 * n_trace_processes:
-            #     print("Waiting for tracing to catch up")
-            #     #initialize time counter
-            #     wait_t0 = time.time()
-            #     while len(ctw.chunknames_written) > len(trace_pool_results) + 2 * n_trace_processes:
-            #         # print dot every second
-            #         if time.time() - wait_t0 > 1:
-            #             print('.', end='')
-                # print("")
-            while len(ctw.chunknames_written) > len(trace_pool_results) + 2 * n_trace_processes:
-                print("waiting for tracing to catch up")
-                time.sleep(5)
-
-        ## Wait for trace to complete
-        if verbose:
-            print("done with reading and writing, just waiting for tracing")
-        # Tell it no more jobs, so close when done
-        trace_pool.close()
-
-        # Wait for everything to finish
-        trace_pool.join()
-
-        ## Error check the tifs that were processed
-        # Get the tifs we wrote, and the tifs we trace
-        written_chunks = sorted(ctw.chunknames_written)
-        traced_filenames = sorted([
-            res['video_filename'] for res in trace_pool_results])
-
-        # Check that they are the same
-        if not np.all(np.array(written_chunks) == np.array(traced_filenames)):
-            raise ValueError("Not all chunks were traced")
-
-        ## Extract the chunk numbers from the filenames
-        # The tiffs have been written, figure out which they are
-        split_traced_filenames = [os.path.split(fn)[1] for fn in traced_filenames]
-        
-        # Replace the format specifier with (\d+) in the pattern
-        mod_chunk_name_pattern = re.sub(r'%\d+d', r'(\\d+)', chunk_name_pattern)
-        mod_chunk_name_pattern = '^' + mod_chunk_name_pattern + '$'
-        tif_file_number_strings = wwutils.misc.apply_and_filter_by_regex(
-            mod_chunk_name_pattern, split_traced_filenames, sort=False)
-
-        # Replace the format specifier with %s in the pattern
-        tif_chunk_name_pattern = re.sub(r'%\d+d', r'%s', chunk_name_pattern)
-        tif_full_filenames = [
-            os.path.join(tiffs_to_trace_directory, tif_chunk_name_pattern % fns)
-            for fns in tif_file_number_strings]
-        
-        tif_file_numbers = list(map(int, tif_file_number_strings))
-        tif_ordering = np.argsort(tif_file_numbers)
-        tif_sorted_filenames = np.array(tif_full_filenames)[
-            tif_ordering]
-        tif_sorted_file_numbers = np.array(tif_file_numbers)[
-            tif_ordering]
-        
-        sorted_file_numbers = tif_sorted_file_numbers
-        sorted_filenames = tif_sorted_filenames
-
-    # stitch
-    if not skip_stitch:
-        print("Stitching")
-        zobj = list(zip(sorted_file_numbers, sorted_filenames))
-        # if output_filename contains h5
-        if output_filename.endswith('.hdf5'):
-            for chunk_start, chunk_name in zobj:
-                # Append each chunk to the hdf5 file
-                if chunk_name.endswith('.tif'):
-                    fn = wwutils.utils.FileNamer.from_tiff_stack(chunk_name)
-                elif chunk_name.endswith('.whiskers'):
-                    fn = wwutils.utils.FileNamer.from_whiskers(chunk_name)
-                append_whiskers_to_hdf5(
-                    whisk_filename=fn.whiskers,
-                    measurements_filename = fn.measurements,
-                    h5_filename=output_filename,
-                    chunk_start=chunk_start,
-                    summary_only=summary_only,
-                    face_side=face)
-                
-        elif output_filename.endswith('.parquet'):
-            try:
-                # Check whether the stitched file has been successfully created previously
-                if not os.path.exists(output_filename):
-                    # Then check if the chunk files need to be created first
-                    if 'temp_dir' not in locals() or \
-                        temp_dir is None or \
-                        not os.path.exists(temp_dir) or \
-                        len(os.listdir(temp_dir)) != len(sorted_filenames):
-                            # Need to create the chunk files first before merging
-                            temp_dir = tempfile.mkdtemp()
-                            # Prepare the arguments for parallel processing
-                            args_list = [(whiskers_file, extract_chunk_number(whiskers_file), face, temp_dir) for whiskers_file in sorted_filenames]
-                            with mp.Pool(n_trace_processes) as parquet_pool:
-                                parquet_pool.starmap(
-                                    process_and_write_parquet, 
-                                    args_list
-                                )
-                            
-                            # for whiskers_file in sorted_filenames:
-                            #     chunk_start = extract_chunk_number(whiskers_file)
-                            #     process_and_write_parquet(whiskers_file, chunk_start=chunk_start, face_side=face, temp_dir=temp_dir)
-                            
-                    # Merge all chunk Parquet files into the final output Parquet file
-                    merge_parquet_files(temp_dir, output_filename)
-            finally:
-                # Clean up the temporary directory if it exists
-                if 'temp_dir' in locals() and temp_dir is not None:
-                    shutil.rmtree(temp_dir)
-                
-        elif output_filename.endswith('.zarr'):
-            # Direct parallel writing to Zarr with consolidated metadata
-            with mp.Pool() as pool:
-                pool.starmap(
-                    process_and_write_zarr, 
-                    [(fn, chunk_start, output_filename) for chunk_start, fn in zobj]
-                )
-            # Consolidate Zarr metadata
-            consolidate_zarr_metadata(output_filename)
-        # elif output_filename.endswith('.zarr'): 
-        #     append_whiskers_to_zarr(
-        #         whisk_filename=fn.whiskers,
-        #         measurements_filename=fn.measurements,
-        #         zarr_filename=output_filename,
-        #         chunk_start=chunk_start,
-        #         summary_only=summary_only)
-
-    # Finalize writers
-    ctw.close()
-    if ffw is not None:
-        ff_stdout, ff_stderr = ffw.close()
-    else:
-        ff_stdout, ff_stderr = None, None
-
-    # Also write timestamps as numpy file
-    if hasattr(input_reader, 'timestamps') and timestamps_filename is not None:
-        timestamps = np.concatenate(input_reader.timestamps)
-        assert len(timestamps) >= ctw.frames_written
-        np.save(timestamps_filename, timestamps[:ctw.frames_written])
-
-    return {'trace_pool_results': trace_pool_results,
-        'monitor_ff_stdout': ff_stdout,
-        'monitor_ff_stderr': ff_stderr,
-        'tif_sorted_file_numbers': sorted_file_numbers,
-        'tif_sorted_filenames': sorted_filenames,
-        }
-
-def compress_pf_to_video(input_reader, chunk_size=200, stop_after_frame=None,
-    timestamps_filename=None, monitor_video=None, monitor_video_kwargs=None,
-    write_monitor_ffmpeg_stderr_to_screen=False, frame_func=None, verbose=True,
-    ):
-    """Read modulated data and compress to video
-
-    Adapted from interleaved_reading_and_tracing
-
-    input_reader : typically a PFReader
-    chunk_size : frames per chunk
-    stop_after_frame : break early, for debugging
-    timestamps_filename : Where to store the timestamps
-        Only valid for PFReader input_reader
-    monitor_video : filename for a monitor video
-        If None, no monitor video will be written
-    monitor_video_kwargs : kwargs to pass to FFmpegWriter for monitor
-        If None, the default is {'qp': 15} for a high-fidelity compression
-        that is still ~6x smaller than lossless.
-    write_monitor_ffmpeg_stderr_to_screen : whether to display
-        output from ffmpeg writing instance
-    frame_func : function to apply to each frame
-        If 'invert', will apply 255 - frame
-    verbose : verbose
-
-    Returns: dict
-        monitor_ff_stderr, monitor_ff_stdout : results from monitor
-            video ffmpeg instance
-    """
-    ## Set up kwargs
-    if monitor_video_kwargs is None:
-        monitor_video_kwargs = {'qp': 15}
-
-    if frame_func == 'invert':
-        frame_func = lambda frame: 255 - frame
-
-    ## Initialize readers and writers
-    if verbose:
-        print("initalizing readers and writers")
-
-    # FFmpeg writer is initalized after first frame
-    ffw = None
-
-    ## Iterate over chunks
-    out_of_frames = False
-    nframe = 0
-    nframes_written = 0
-
-    # Init the iterator outside of the loop so that it persists
-    iter_obj = input_reader.iter_frames()
-
-    while not out_of_frames:
-        # Get a chunk of frames
-        if verbose:
-            print("loading chunk of frames starting with", nframe)
-        chunk_of_frames = []
-        for frame in iter_obj:
-            if frame_func is not None:
-                frame = frame_func(frame)
-            chunk_of_frames.append(frame)
-            nframe = nframe + 1
-            if stop_after_frame is not None and nframe >= stop_after_frame:
-                break
-            if len(chunk_of_frames) == chunk_size:
-                break
-
-        # Check if we ran out
-        if len(chunk_of_frames) != chunk_size:
-            out_of_frames = True
-
-        ## Start monitor encode
-        # This is also synchronous, otherwise the input buffer might fill up
-        if monitor_video is not None:
-            if ffw is None:
-                ffw = WhiskiWrap.FFmpegWriter(monitor_video,
-                    frame_width=frame.shape[1], frame_height=frame.shape[0],
-                    write_stderr_to_screen=write_monitor_ffmpeg_stderr_to_screen,
-                    **monitor_video_kwargs)
-            for frame in chunk_of_frames:
-                ffw.write(frame)
-                nframes_written = nframes_written + 1
-
-    # Finalize writers
-    if ffw is not None:
-        ff_stdout, ff_stderr = ffw.close()
-    else:
-        ff_stdout, ff_stderr = None, None
-
-    # Also write timestamps as numpy file
-    if hasattr(input_reader, 'timestamps') and timestamps_filename is not None:
-        timestamps = np.concatenate(input_reader.timestamps)
-
-        # These assertions only make sense if we wrote the whole file
-        if stop_after_frame is None:
-            assert len(timestamps) == nframes_written
-            assert nframes_written == nframe
-
-        # Save timestamps
-        np.save(timestamps_filename, timestamps)
-
-    return {
-        'monitor_ff_stdout': ff_stdout,
-        'monitor_ff_stderr': ff_stderr,
-    }
-
-def measure_chunk_star(args):
-    return measure_chunk(*args)
-
-def read_whiskers_hdf5_summary(filename):
-    """Reads and returns the `summary` table in an HDF5 file"""
-    with tables.open_file(filename) as fi:
-        # Check whether the summary table exists, or updated_summary
-        if '/summary' in fi:
-            summary = pd.DataFrame.from_records(fi.root.summary.read())
-        elif '/updated_summary' in fi:
-            try:
-                # Assuming a group
-                summary = pd.DataFrame.from_records(fi.root.updated_summary.read())
-            except:
-                # Then assuming a table
-                table_node = fi.get_node('/updated_summary/table')
-                summary = pd.DataFrame.from_records(table_node.read())
-        else:
-            raise ValueError("no summary table found")
-
-    return summary
-
-def read_whiskers_measurements(filenames):
-    # Loads all the whiskers/measurements files and returns an aggregated dataframe
-
-    # Check that it is a list
-    if not isinstance(filenames, list):
-        filenames = [filenames]
-
-    # Load the measurements
-    all_measurements = []
-    for filename in filenames:
-        #  strip the extension and replace with .whiskers
-        if not filename.endswith('.whiskers'):
-            filename = filename.replace(filename.split('.')[-1], 'whiskers')
-        # get the whisker data
-        measurements = WhiskiWrap.read_whisker_data(filename, 'df')
-        # append the dataframe to the list
-        all_measurements.append(measurements)
-
-    # Concatenate the dataframes
-    all_measurements = pd.concat(all_measurements)
-
-    return all_measurements
-
-class PFReader:
-    """Reads photonfocus modulated data stored in matlab files"""
-    def __init__(self, input_directory, n_threads=4, verbose=True,
-        error_on_unsorted_filetimes=True):
-        """Initialize a new reader.
-
-        input_directory : where the mat files are
-            They are assumed to have a format like img10.mat, etc.
-            They should contain variables called 'img' (a 4d array of
-            modulated frames) and 't' (timestamps of each frame).
-        n_threads : sent to pfDoubleRate_SetNrOfThreads
-
-        error_on_unsorted_filetimes : bool
-            Whether to raise an error if the modification times of the
-            matfiles are not in sorted order, which typically happens if
-            something has gone wrong (but could just be that the file times
-            weren't preserved)
-        """
-        self.input_directory = input_directory
-        self.verbose = verbose
-
-        ## Load the libraries
-        # boost_thread needs boost_system
-        # I think it used to be able to find boost_system without this line
-        libboost_system = ctypes.cdll.LoadLibrary(
-            '/usr/local/lib/libboost_system.so.1.50.0')
-
-        # Load boost_thread
-        libboost_thread = ctypes.cdll.LoadLibrary(
-            '/usr/local/lib/libboost_thread.so')
-
-        # Load the pf_lib (which requires boost_thread)
-        self.pf_lib = ctypes.cdll.LoadLibrary(LIB_DOUBLERATE)
-        self.demod_func = self.pf_lib['pfDoubleRate_DeModulateImage']
-
-        # Set the number of threads
-        self.pf_lib['pfDoubleRate_SetNrOfThreads'](n_threads)
-
-        # Find all the imgN.mat files in the input directory
-        self.matfile_number_strings = wwutils.misc.apply_and_filter_by_regex(
-            '^img(\d+)\.mat$', os.listdir(self.input_directory), sort=False)
-        self.matfile_names = [
-            os.path.join(self.input_directory, 'img%s.mat' % fns)
-            for fns in self.matfile_number_strings]
-        self.matfile_numbers = list(map(int, self.matfile_number_strings))
-        self.matfile_ordering = np.argsort(self.matfile_numbers)
-
-        # Sort the names and numbers
-        self.sorted_matfile_names = np.array(self.matfile_names)[
-            self.matfile_ordering]
-        self.sorted_matfile_numbers = np.array(self.matfile_numbers)[
-            self.matfile_ordering]
-
-        # Error check the file times
-        filetimes = np.array([
-            wwutils.misc.get_file_time(filename)
-            for filename in self.sorted_matfile_names])
-        if (np.diff(filetimes) < 0).any():
-            if error_on_unsorted_filetimes:
-                raise IOError("unsorted matfiles")
-            else:
-                print("warning: unsorted matfiles")
-
-        # Create variable to store timestamps
-        self.timestamps = []
-        self.n_frames_read = 0
-
-        # Set these values once they are read from the file
-        self.frame_height = None
-        self.frame_width = None
-
-    def iter_frames(self):
-        """Yields frames as they are read and demodulated.
-
-        Iterates through the matfiles in order, demodulates each frame,
-        and yields them one at a time.
-
-        The chunk of timestamps from each matfile is appended to the list
-        self.timestamps, so that self.timestamps is a list of arrays. There
-        will be more timestamps than read frames until the end of the chunk.
-
-        Also sets self.frame_height and self.frame_width and checks that
-        they are consistent over the session.
-        """
-        # Iterate through matfiles and load
-        for matfile_name in self.sorted_matfile_names:
-            if self.verbose:
-                print(("loading %s" % matfile_name))
-
-            # Load the raw data
-            # This is the slowest step
-            matfile_load = scipy.io.loadmat(matfile_name)
-            matfile_t = matfile_load['t'].flatten()
-            matfile_modulated_data = matfile_load['img'].squeeze()
-            assert matfile_modulated_data.ndim == 3 # height, width, time
-
-            # Append the timestamps
-            self.timestamps.append(matfile_t)
-
-            # Extract shape
-            n_frames = len(matfile_t)
-            assert matfile_modulated_data.shape[-1] == n_frames
-            modulated_frame_width = matfile_modulated_data.shape[1]
-            frame_height = matfile_modulated_data.shape[0]
-
-            if self.verbose:
-                print(("loaded %d modulated frames @ %dx%d" % (n_frames,
-                    modulated_frame_width, frame_height)))
-
-            # Find the demodulated width
-            # This can be done by pfDoubleRate_GetDeModulatedWidth
-            # but I don't understand what it's doing.
-            if modulated_frame_width == 416:
-                demodulated_frame_width = 800
-            elif modulated_frame_width == 332:
-                demodulated_frame_width = 640
-            else:
-                raise ValueError("unknown modulated width: %d" %
-                    modulated_frame_width)
-
-            # Store the frame sizes as necessary
-            if self.frame_width is None:
-                self.frame_width = demodulated_frame_width
-            if self.frame_width != demodulated_frame_width:
-                raise ValueError("inconsistent frame widths")
-            if self.frame_height is None:
-                self.frame_height = frame_height
-            if self.frame_height != frame_height:
-                raise ValueError("inconsistent frame heights")
-
-            # Create a buffer for the result of each frame
-            demodulated_frame_buffer = ctypes.create_string_buffer(
-                frame_height * demodulated_frame_width)
-
-            # Iterate over frames
-            for n_frame in range(n_frames):
-                if self.verbose and np.mod(n_frame, 200) == 0:
-                    print(("iterator has reached frame %d" % n_frame))
-
-                # Convert image to char array
-                # Ideally we would use a buffer here instead of a copy in order
-                # to save time. But Matlab data comes back in Fortran order
-                # instead of C order, so this is not possible.
-                frame_charry = ctypes.c_char_p(
-                    matfile_modulated_data[:, :, n_frame].tobytes())
-                #~ frame_charry = ctypes.c_char_p(
-                    #~ bytes(matfile_modulated_data[:, :, n_frame].data))
-
-                # Run
-                self.demod_func(demodulated_frame_buffer, frame_charry,
-                    demodulated_frame_width, frame_height,
-                    modulated_frame_width)
-
-                # Extract the result from the buffer
-                demodulated_frame = np.fromstring(demodulated_frame_buffer,
-                    dtype=np.uint8).reshape(
-                    frame_height, demodulated_frame_width)
-
-                self.n_frames_read = self.n_frames_read + 1
-
-                yield demodulated_frame
-
-        if self.verbose:
-            print("iterator is empty")
-
-    def close(self):
-        """Currently does nothing"""
-        pass
-
-    def isclosed(self):
-        return True
-
-class ChunkedTiffWriter:
-    """Writes frames to a series of tiff stacks"""
-    def __init__(self, output_directory, chunk_size=200, chunk_name_pattern='chunk%08d.tif', compress=False):
-        """Initialize a new chunked tiff writer.
-
-        output_directory : where to write the chunks
-        chunk_size : frames per chunk
-        chunk_name_pattern : how to name the chunk, using the number of
-            the first frame in it
-        """
-        self.output_directory = output_directory
-        self.chunk_size = chunk_size
-        self.chunk_name_pattern = chunk_name_pattern
-        self.compress = compress
-
-        # Initialize counters so we know what frame and chunk we're on
-        self.frames_written = 0
-        self.frame_buffer = []
-        self.chunknames_written = []
-
-    def write(self, frame):
-        """Buffered write frame to tiff stacks"""
-        # Append to buffer
-        self.frame_buffer.append(frame)
-
-        # Write chunk if buffer is full
-        if len(self.frame_buffer) == self.chunk_size:
-            self._write_chunk()
-
-    def _write_chunk(self):
-        if len(self.frame_buffer) != 0:
-            # Form the chunk
-            chunk = np.array(self.frame_buffer)
-
-            # Name it
-            chunkname = os.path.join(self.output_directory,
-                self.chunk_name_pattern % self.frames_written)
-
-            # Write it
-            if self.compress:
-                if hasattr(tifffile, 'imwrite'):
-                    tifffile.imwrite(chunkname, chunk, compression='lzw')
-                else:
-                    tifffile.imsave(chunkname, chunk, compression='lzw')
-            else:
-                if hasattr(tifffile, 'imwrite'):
-                    tifffile.imwrite(chunkname, chunk, compression=None)
-                else:
-                    tifffile.imsave(chunkname, chunk, compression=False)
-
-            # Update the counter
-            self.frames_written += len(self.frame_buffer)
-            self.frame_buffer = []
-
-            # Update the list of written chunks
-            self.chunknames_written.append(chunkname)
-
-    def count_unwritten_frames(self):
-        """Returns the number of buffered, unwritten frames"""
-        return len(self.frame_buffer)
-
-    def close(self):
-        """Finish writing any final unfinished chunk"""
-        self._write_chunk()
-
-class FFmpegReader:
-    """Reads frames from a video file using ffmpeg process"""
-    def __init__(self, input_filename, pix_fmt='gray', bufsize=10**9,
-        duration=None, start_frame_time=None, start_frame_number=None, crop=None,
-        write_stderr_to_screen=False, vsync='drop'):
-        """Initialize a new reader
-
-        input_filename : name of file
-        pix_fmt : used to format the raw data coming from ffmpeg into
-            a numpy array
-        bufsize : probably not necessary because we read one frame at a time
-        duration : duration of video to read (-t parameter)
-        start_frame_time, start_frame_number : -ss parameter
-            Parsed using video.ffmpeg_frame_string
-        crop : crop the video using the ffmpeg crop filter. Format is width:height:x:y
-        write_stderr_to_screen : if True, writes to screen, otherwise to
-            /dev/null
-        """
-        self.input_filename = input_filename
-
-        # Get params
-        self.frame_width, self.frame_height, self.frame_rate = \
-            video.get_video_params(input_filename, crop=crop)
-
-        # Set up pix_fmt
-        if pix_fmt == 'gray':
-            self.bytes_per_pixel = 1
-        elif pix_fmt == 'rgb24':
-            self.bytes_per_pixel = 3
-        else:
-            raise ValueError("can't handle pix_fmt:", pix_fmt)
-        self.read_size_per_frame = self.bytes_per_pixel * \
-            self.frame_width * self.frame_height
-
-        # Create the command
-        command = ['ffmpeg']
-
-        # Add ss string
-        if start_frame_time is not None or start_frame_number is not None:
-            ss_string = video.ffmpeg_frame_string(input_filename,
-                frame_time=start_frame_time, frame_number=start_frame_number)
-            command += [
-                '-ss', ss_string]
-
-        # Add input file
-        command += [
-            '-i', input_filename,
-            '-vsync', vsync,
-            '-f', 'image2pipe',
-            '-pix_fmt', pix_fmt]
-        
-        # Add crop string. The crop argument is a list
-        if crop is not None:
-            self.crop = crop
-        else:
-            self.crop = None
-        #     command += [
-        #         '-vf', 'crop=%d:%d:%d:%d' % tuple(crop)]
-
-        # Add duration string
-        if duration is not None:
-            command += [
-                '-t', str(duration),]
-
-        # Add vcodec for pipe
-        command += [
-            '-vcodec', 'rawvideo', '-']
-
-        # To store result
-        self.n_frames_read = 0
-
-        # stderr
-        if write_stderr_to_screen:
-            stderr = None
-        else:
-            stderr = open(os.devnull, 'w')
-
-        # Init the pipe
-        # We set stderr to null so it doesn't fill up screen or buffers
-        # And we set stdin to PIPE to keep it from breaking our STDIN
-        self.ffmpeg_proc = subprocess.Popen(command,
-            stdin=subprocess.PIPE,
-            stdout=subprocess.PIPE, stderr=stderr,
-            bufsize=bufsize)
-
-    def iter_frames(self):
-        """Yields one frame at a time
-
-        When done: terminates ffmpeg process, and stores any remaining
-        results in self.leftover_bytes and self.stdout and self.stderr
-
-        It might be worth writing this as a chunked reader if this is too
-        slow. Also we need to be able to seek through the file.
-        """
-        # Read this_chunk, or as much as we can
-        while(True):
-            raw_image = self.ffmpeg_proc.stdout.read(self.read_size_per_frame)
-
-            # if self.crop is not None:
-                # import matplotlib.pyplot as plt
-                # # plot raw image, full and cropped (cropp coordinates format is width:height:x:y)
-                # raw_image_full = np.frombuffer(raw_image, dtype='uint8').reshape((self.frame_height, self.frame_width, self.bytes_per_pixel))
-                # raw_image_cropped = raw_image_full[self.crop[3]:self.crop[3]+self.crop[1], self.crop[2]:self.crop[2]+self.crop[0], :]
-                # plt.figure()
-                # plt.subplot(1,2,1)
-                # plt.imshow(raw_image_full)
-                # plt.subplot(1,2,2)
-                # plt.imshow(raw_image_cropped)
-                # plt.show()
-
-            # check if we ran out of frames
-            if len(raw_image) != self.read_size_per_frame:
-                self.leftover_bytes = raw_image
-                self.close()
-                return
-
-            # Convert to array, flatten and squeeze (and crop if required, although it is slightly faster to crop in the calling function with frame_func)
-            if self.crop is None:
-                frame = np.frombuffer(raw_image, dtype='uint8').reshape((self.frame_height, self.frame_width, self.bytes_per_pixel)).squeeze()
-            else:
-                frame = np.frombuffer(raw_image, dtype='uint8').reshape((self.frame_height, self.frame_width, self.bytes_per_pixel)).squeeze()[self.crop[3]:self.crop[3]+self.crop[1], self.crop[2]:self.crop[2]+self.crop[0]]
-
-            # Update
-            self.n_frames_read = self.n_frames_read + 1
-
-            # Yield
-            yield frame
-
-    def close(self):
-        """Closes the process"""
-        # Need to terminate in case there is more data but we don't
-        # care about it
-        # But if it's already terminated, don't try to terminate again
-        if self.ffmpeg_proc.returncode is None:
-            self.ffmpeg_proc.terminate()
-
-            # Extract the leftover bits
-            self.stdout, self.stderr = self.ffmpeg_proc.communicate()
-
-        return self.ffmpeg_proc.returncode
-
-    def isclosed(self):
-        if hasattr(self.ffmpeg_proc, 'returncode'):
-            return self.ffmpeg_proc.returncode is not None
-        else:
-            # Never even ran? I guess this counts as closed.
-            return True
-
-class FFmpegWriter:
-    """Writes frames to an ffmpeg compression process"""
-    def __init__(self, output_filename, frame_width, frame_height,
-        output_fps=30, vcodec='libx264', qp=15, preset='medium',
-        input_pix_fmt='gray', output_pix_fmt='yuv420p',
-        write_stderr_to_screen=False):
-        """Initialize the ffmpeg writer
-
-        output_filename : name of output file
-        frame_width, frame_height : Used to inform ffmpeg how to interpret
-            the data coming in the stdin pipe
-        output_fps : frame rate
-        input_pix_fmt : Tell ffmpeg how to interpret the raw data on the pipe
-            This should match the output generated by frame.tostring()
-        output_pix_fmt : pix_fmt of the output
-        crf : quality. 0 means lossless
-        preset : speed/compression tradeoff
-        write_stderr_to_screen :
-            If True, writes ffmpeg's updates to screen
-            If False, writes to /dev/null
-
-        With old versions of ffmpeg (jon-severinsson) I was not able to get
-        truly lossless encoding with libx264. It was clamping the luminances to
-        16..235. Some weird YUV conversion?
-        '-vf', 'scale=in_range=full:out_range=full' seems to help with this
-        In any case it works with new ffmpeg. Also the codec ffv1 will work
-        but is slightly larger filesize.
-        """
-        # Open an ffmpeg process
-        cmdstring = ('ffmpeg',
-            '-y', '-r', '%d' % output_fps,
-            '-s', '%dx%d' % (frame_width, frame_height), # size of image string
-            '-pix_fmt', input_pix_fmt,
-            '-f', 'rawvideo',  '-i', '-', # raw video from the pipe
-            '-pix_fmt', output_pix_fmt,
-            '-vcodec', vcodec,
-            '-qp', str(qp),
-            '-preset', preset,
-            output_filename) # output encoding
-
-        if write_stderr_to_screen:
-            self.ffmpeg_proc = subprocess.Popen(cmdstring, stdin=subprocess.PIPE,
-                stdout=subprocess.PIPE)
-        else:
-            self.ffmpeg_proc = subprocess.Popen(cmdstring, stdin=subprocess.PIPE,
-                stdout=subprocess.PIPE, stderr=open('/dev/null', 'w'))
-
-    def write(self, frame):
-        """Write a frame to the ffmpeg process"""
-        self.ffmpeg_proc.stdin.write(frame.tostring())
-
-    def write_bytes(self, bytestring):
-        self.ffmpeg_proc.stdin.write(bytestring)
-
-    def close(self):
-        """Closes the ffmpeg process and returns stdout, stderr"""
-        return self.ffmpeg_proc.communicate()
-
-
-def interleaved_trace_and_measure(input_reader, tiffs_to_trace_directory,
-    sensitive=False,
-    chunk_size=200, chunk_name_pattern='chunk%08d.tif',
-    stop_after_frame=None, delete_tiffs=True,
-    timestamps_filename=None, monitor_video=None,
-    monitor_video_kwargs=None, write_monitor_ffmpeg_stderr_to_screen=False,
-    output_filename=None, frame_func=None,
-    n_trace_processes=4, expectedrows=1000000,
-    verbose=True, skip_stitch=False, classify=None,
-    summary_only=False, skip_existing=False, convert_chunks=False
-    ):
-    """Read, write, trace and measure each chunk, one at a time.
-
-    This function is designed for single-side whisker tracking where the full video
-    frame is processed without splitting. It combines tracing and measurement in
-    a single pipeline with Parquet output support.
-
-    input_reader : Typically a PFReader or FFmpegReader
-    tiffs_to_trace_directory : Location to write the tiffs
-    sensitive: if False, use default. If True, lower MIN_SIGNAL
-    chunk_size : frames per chunk
-    chunk_name_pattern : how to name them
-    stop_after_frame : break early, for debugging
-    delete_tiffs : whether to delete tiffs after done tracing
-    timestamps_filename : Where to store the timestamps
-        Only valid for PFReader input_reader
-    monitor_video : filename for a monitor video
-        If None, no monitor video will be written
-    monitor_video_kwargs : kwargs to pass to FFmpegWriter for monitor
-    write_monitor_ffmpeg_stderr_to_screen : whether to display
-        output from ffmpeg writing instance
-    output_filename : Parquet file to write whisker data to
-    frame_func : function to apply to each frame
-        If 'invert', will apply 255 - frame
-    n_trace_processes : number of simultaneous trace processes
-    expectedrows : expected number of whisker segments (for optimization)
-    verbose : whether to print progress information
-    skip_stitch : if True, don't stitch results together
-    classify : dictionary of classify arguments for whisker classification
-    summary_only : if True, only save summary data (not pixel coordinates)
-    skip_existing : if True, skip processing if output file already exists
-    convert_chunks : if True, convert individual chunks to desired format
-    """
-    if skip_existing and output_filename and os.path.exists(output_filename):
-        if verbose:
-            print(f"Output file {output_filename} already exists, skipping...")
-        return {'output_filename': output_filename}
-
-    # Set up parameters file copying
-    copy_parameters_files(tiffs_to_trace_directory, sensitive=sensitive)
-    
-    # Initialize output tracking
-    result_dict = {}
-    
-    # Create temporary directory for chunk processing
-    temp_dir = os.path.join(tiffs_to_trace_directory, 'temp_chunks')
-    os.makedirs(temp_dir, exist_ok=True)
-    
-    try:
-        # Write video as chunked tiffs and process them
-        ctw = write_video_as_chunked_tiffs(
-            input_reader, 
-            tiffs_to_trace_directory,
-            chunk_size=chunk_size, 
-            chunk_name_pattern=chunk_name_pattern,
-            stop_after_frame=stop_after_frame, 
-            monitor_video=monitor_video,
-            timestamps_filename=timestamps_filename,
-            monitor_video_kwargs=monitor_video_kwargs
-        )
-        
-        # Get list of tiff files to process
-        tif_filenames = ctw.chunknames_written
-        
-        if verbose:
-            print(f"Processing {len(tif_filenames)} chunks with {n_trace_processes} processes")
-        
-        # Process chunks in parallel
-        pool = mp.Pool(n_trace_processes)
-        
-        # Keep track of results
-        trace_results = []
-        def log_result(result):
-            print("Result logged:", result)
-            trace_results.append(result)
-        
-        # Process all chunks
-        if verbose:
-            print("Tracing and measuring chunks...")
-            
-        # Start async jobs for each chunk
-        convert_chunks_to = 'parquet' if convert_chunks else None
-        for tif_filename in tif_filenames:
-            pool.apply_async(trace_and_measure_chunk, 
-                           args=(tif_filename, delete_tiffs, 'right', classify, 
-                                temp_dir if convert_chunks else None, convert_chunks_to),
-                           callback=log_result)
-        
-        # Wait for all jobs to complete
-        pool.close()
-        pool.join()
-        
-        if not skip_stitch and output_filename:
-            if verbose:
-                print("Stitching results together...")
-            
-            if convert_chunks:
-                # Merge parquet files from temp directory
-                merge_parquet_files(temp_dir, output_filename)
-            else:
-                # Convert and stitch HDF5 files to Parquet
-                stitch_h5_to_parquet(tif_filenames, output_filename, face_side='single')
-        
-        # Clean up temporary directory
-        if os.path.exists(temp_dir):
-            import shutil
-            shutil.rmtree(temp_dir)
-        
-        result_dict['output_filename'] = output_filename
-        result_dict['trace_results'] = trace_results
-        result_dict['n_chunks'] = len(tif_filenames)
-        
-        if verbose:
-            print(f"Completed processing. Output saved to {output_filename}")
-        
-        return result_dict
-        
-    except Exception as e:
-        # Clean up on error
-        if os.path.exists(temp_dir):
-            import shutil
-            shutil.rmtree(temp_dir)
-        raise e
-
-def stitch_h5_to_parquet(tif_filenames, output_filename, face_side='single'):
-    """
-    Stitch together HDF5 files from individual chunks into a single Parquet file.
-    
-    This is a helper function for interleaved_trace_and_measure when convert_chunks=False.
-    """
-    all_data = []
-    
-    for chunk_idx, tif_filename in enumerate(tif_filenames):
-        h5_filename = tif_filename.replace('.tif', '.hdf5')
-        if os.path.exists(h5_filename):
-            chunk_start = chunk_idx * 200  # Default chunk size
-            append_whiskers_to_parquet(
-                whisk_filename=tif_filename.replace('.tif', '.whiskers'),
-                measurements_filename=tif_filename.replace('.tif', '.measurements'),
-                parquet_filename=output_filename,
-                chunk_start=chunk_start,
-                summary_only=False,
-                face_side=face_side
-            )
-            # Clean up individual HDF5 file
-            os.remove(h5_filename)
+\nfrom .io import PFReader, ChunkedTiffWriter, FFmpegReader, FFmpegWriter
+from .pipeline import pipeline_trace, write_video_as_chunked_tiffs, trace_chunked_tiffs, interleaved_read_trace_and_measure, interleaved_split_trace_and_measure, interleaved_trace_and_measure, compress_pf_to_video, measure_chunk_star, read_whiskers_hdf5_summary, read_whiskers_measurements, stitch_h5_to_parquet
